[#88] StockDetailController 추가
- getStockDetails 메서드 추가
