+++ src/main/java/org/kiworkshop/snowball/stockdetail/entity/StockDetailRepository.java
@@ -1,0 +1,6 @@
+package org.kiworkshop.snowball.stockdetail.entity;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StockDetailRepository extends JpaRepository<StockDetail, Long> {
+}
